+++ NEWS.md
@@ -3,6 +3,7 @@
 New features:
 * Improved column alignment capabilities (via `col_style_plan`). Alignment options now fall into two types: character (type = "char") and positional (type = "pos"). Positional alignment is new and allows for aligning across multiple positions. 
 * Add `page_plan` for splitting tables across multiple pages
+* Add ability to add a group/label header via the `col_plan`
 
 Bug fixes:
 * `frmt_combine` no longer throws error if group variable is named "var"